+++ modules/client/src/client-common/interfaces/common.ts
@@ -3,6 +3,7 @@ export const SupportedNetworksArray = [
   "goerli",
   "matic",
   "maticmum",
+  "mvm",
 ] as const;
 export type SupportedNetworks = typeof SupportedNetworksArray[number];
 export type NetworkDeployment = {