+++ modules/client/src/internal/client/decoding.ts
@@ -19,7 +19,7 @@ import { AVAILABLE_FUNCTION_SIGNATURES } from "../constants";
 import {
   DAO__factory,
   PluginSetupProcessor__factory,
-} from "@aragon/osx-ethers";
+} from "@mixinao/osx-ethers";
 import {
   applyInstallatonParamsFromContract,
   permissionParamsFromContract,